python/visualizer/app/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
yt-dlp==2023.10.13
openai-whisper==20231117
pydub==0.25.1
Werkzeug==2.3.7